requirements.txt
asgiref==3.2.10
attrs==19.3.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==3.0
Django==3.0.8
django-cors-headers==3.4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.10
jellyfish==0.8.2
lib50==2.0.8
pexpect==4.8.0
ptyprocess==0.6.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
submit50==3.0.2
termcolor==1.1.0
urllib3==1.25.10
whitenoise==5.2.0
